% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.star.test.R
\name{f.star.test}
\alias{f.star.test}
\title{f.star.test}
\usage{
f.star.test(means, variances, ns)
}
\arguments{
\item{means}{a (non-empty) numeric vector of the group means}

\item{variances}{a (non-empty) numeric vector of the group variances}

\item{ns}{a (non-empty) numeric vector of sample sizes per group}
}
\value{
\item{statistic}{the value of the adjusted Brown-Forsythe F star statistic}
        \item{p.value}{the p-value for the test}
        \item{est.f.squared}{effect size estimate as in Cohen's f squared}
}
\description{
conduct a Brown-Forsythe F star test
}
\examples{
# a f star test for three-group mean comparison
f.star.test(c(-.2,0,.2),c(1,1,1),c(100,100,100))
f.star.test(c(0,0,1),c(2,1,3),c(100,100,100))
}
